fix: 티저 페이지 임시 url 로 수정
